separate apps to one for 'login' and one for 'list up accounts'
