+++ acclist/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path('', views.index, name='index'),
+    path('<str:user_id>/', views.alllist, name='alllist'),
+]
