+++ accounts/admin.py
@@ -1,3 +1,6 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import User
+
+admin.site.register(User)
+
