+++ accup/urls.py
@@ -18,5 +18,6 @@ from django.urls import include, path
 
 urlpatterns = [
     path('accounts/', include('accounts.urls')),
+    path('acclist/', include('acclist.urls')),
     path('admin/', admin.site.urls),
 ]
